Move SpendsUnconfirmed trait to its own module
